Added code to enable pipelines build with new bootstrap-build-framework.sh script.
Added bootstrap step to all pipelines. This produces the scripts/make directory as an artifact.

Altered the boostrap-build-framework.sh script to tell if it is running under bb pipelines.
If it is running under pipelines, then it need not run its clone_script in a container, as its already the right one (deeson/deployer).

Added a pipelines-boostrap target to Makefile.